+++ openmc/deplete/results.py
